pim-compiler/pim_compiler_chatbot/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0
python-dotenv>=1.0.0
gradio>=4.0.0